functions/get_results/requirements.txt
google-cloud-storage
google-cloud-firestore
google-auth